\name{hleaps-package}
\alias{hleaps-package}
\docType{package}
\title{
Hierarchy Supporting Ordinary Least Squares
}
\description{
Exhaustive search of hierarchy supporting ordinary least squares models
}
\details{
\tabular{ll}{
Package: \tab hleaps\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-22\cr
License:  \tab GPL (>= 2)\cr
}
}
\author{
Mark Banghart <mbanghart@ssc.wisc.edu>, Arun Srinivasan <aasrinivasan@wisc.edu>, and Douglas Bates.

Maintainer: Mark Banghart <mbanghart@ssc.wisc.edu>, Arun Srinivasan <aasrinivasan@wisc.edu>
}